restructured code and fixed compilation errors
