--- Networking/.env
@@ -1,1 +1,0 @@
-NGROK_AUTH = "2YBQtBWMx3e6X4XzzhgJLgqWD2O_5oftTaZ13Ag4Hr1ktHc6e"
