--- RequestResponse/task_response.cpp
@@ -1,1 +1,0 @@
-#include "task_response.h"
