+++ src/Networking/client.cpp
@@ -1,4 +1,4 @@
-#include "client.h"
+#include "../../include/Networking/client.h"
 
 #include <arpa/inet.h>
 #include <netdb.h>
