+++ src/Networking/server.cpp
@@ -1,4 +1,4 @@
-#include "server.h"
+#include "../../include/Networking/server.h"
 
 #include <arpa/inet.h>
 #include <netinet/in.h>
