+++ src/Networking/utility.cpp
