+++ src/Peers/bootstrap_node.cpp
@@ -1,3 +1,3 @@
-#include "bootstrap_node.h"
+#include "../../include/Peers/bootstrap_node.h"
 
 BootstrapNode::BootstrapNode(std::string uuid, unsigned short port) {}
