+++ src/utility.cpp
@@ -1,6 +1,6 @@
 #include <random>
 #include <sstream>
-#include "utility.h"
+#include "../include/utility.h"
 
 std::string uuid::generate_uuid_v4() {
     std::stringstream ss;
